requirements.txt
adjust==0.1.0
aiodns==3.2.0
aiohappyeyeballs==2.4.4
aiohttp==3.10.11
aiosignal==1.3.2
alembic==1.14.1
alpha_vantage==3.0.0
annotated-types==0.7.0
anyio==4.8.0
asgiref==3.8.1
attrs==24.3.0
backtrader==1.9.78.123
bcolz-zipline==1.2.12
beautifulsoup4==4.12.3
binance==0.3
binary==1.0.1
binary-tree==0.0.1
binaryai==0.8.2
blinker==1.9.0
blosc2==3.0.0
boto==2.49.0
Bottleneck==1.4.2
category_encoders==2.8.0
ccxt==4.4.47
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
ClosedAPI==0.3
colorama==0.4.6
compute_oper==0.0.1
config==0.5.1
configparser==7.1.0
ConfigSpace==1.2.1
contourpy==1.3.1
coverage==7.6.12
cryptography==44.0.0
cycler==0.12.1
data==0.4
databricks-api==0.9.0
databricks-cli==0.18.0
dateparser==1.2.0
decorator==5.1.1
Deprecated==1.2.15
Django==5.1.5
djangorestframework==3.15.2
djangorestframework-csv==3.0.2
download==0.3.5
download-yfinance-data==0.0.1
empyrical-reloaded==0.5.11
exchange_calendars==4.8
Flask==3.0.1
Flask-Cors==5.0.0
fonttools==4.55.3
frozendict==2.4.6
frozenlist==1.5.0
funcsigs==1.0.2
gevent==24.11.1
greenlet==3.1.1
h11==0.14.0
h5py==3.12.1
html5lib==1.1
httpcore==1.0.7
httpx==0.25.2
idna==3.10
iniconfig==2.0.0
intervaltree==3.1.0
iso3166==2.1.1
iso4217==1.12.20240625
itsdangerous==2.2.0
Jinja2==3.1.5
joblib==1.4.2
joblibs3==0.1.0
json-logging-py==0.2
kiwisolver==1.4.8
korean-lunar-calendar==0.3.1
logger==1.4
lru-dict==1.3.0
lxml==5.3.0
Mako==1.3.8
MarkupSafe==3.0.2
matplotlib==3.10.0
more-itertools==10.6.0
msgpack==1.1.0
multidict==6.1.0
multipledispatch==1.0.0
multitasking==0.0.11
MyFinances==0.1.1
ndindex==1.9.2
networkx==3.4.2
numexpr==2.10.2
numpy==2.2.1
oauthlib==3.2.2
openai==0.28.0
packaging==24.2
pandas==2.2.3
pandas-datareader==0.10.0
patsy==1.0.1
peewee==3.17.3
pillow==11.1.0
platformdirs==4.3.6
pluggy==1.5.0
propcache==0.2.1
py-cpuinfo==9.0.0
pycares==4.5.0
pycparser==2.22
pycryptodome==3.21.0
pydantic==2.10.5
pydantic_core==2.27.2
pydot==3.0.4
PyJWT==2.10.1
pyluach==2.2.0
pyparsing==3.2.1
pytest==8.3.5
pytest-cov==6.0.0
python-binance==1.0.27
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-interface==1.6.1
pytz==2024.2
RandomForest1==0.1
regex==2024.11.6
requests==2.32.3
requests-file==2.1.0
s3io==0.1.1
schedule==1.2.2
scikit-learn==1.6.1
scikit_learnt==0.1.1
scipy==1.15.1
seaborn==0.13.2
setuptools==75.8.0
six==1.17.0
sklearn-pandas==2.2.0
sklearn2==0.0.13
sklearn3==0.0.1
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
SQLAlchemy==2.0.37
sqlparse==0.5.3
statsmodels==0.14.4
ta==0.11.0
ta-lib==0.6.1
tables==3.10.2
tablib==3.8.0
tabulate==0.9.0
threadpoolctl==3.5.0
toolz==1.0.0
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2024.2
tzlocal==5.2
unicodecsv==0.14.1
urllib3==2.3.0
webencodings==0.5.1
websocket-client==1.8.0
websocket-client-py3==0.15.0
websocket-client-temp==0.57.1
websocket-client2==1.0
websockets==14.1
Werkzeug==3.1.3
wrapt==1.17.2
yarl==1.18.3
yfinance==0.2.51
yfinanceapi==1.0.2
zipline-reloaded==3.1
zipline.py==0.20.0
ziplineio==0.1.16
zope.event==5.0
zope.interface==7.2
